+++ src/Control/Lens/Type.hs
@@ -61,8 +61,6 @@ module Control.Lens.Type
   , (%%~)
   , (%%=)
 
-  , resultAt
-
   -- * Lateral Composition
   , choosing
   , chosen
@@ -263,18 +261,6 @@ l %%= f = do
 -------------------------------------------------------------------------------
 -- Common Lenses
 -------------------------------------------------------------------------------
-
--- | This lens can be used to change the result of a function but only where
--- the arguments match the key given.
---
--- >>> let f = (+1) % resultAt 3 .~ 8 in (f 2, f 3)
--- (3,8)
-resultAt :: Eq e => e -> Simple Lens (e -> a) a
-resultAt e afa ea = go <$> afa a where
-  a = ea e
-  go a' e' | e == e'   = a'
-           | otherwise = ea e'
-{-# INLINE resultAt #-}
 
 -- | Merge two lenses, getters, setters, folds or traversals.
 --