+++ CHANGELOG.markdown
@@ -13,6 +13,7 @@
 * Added `beside` to `Control.Lens.Traversal`.
 * Removed the operators from `Data.List.Lens`, they broke the overall pattern of the rest of the API, and were terrible clutter.
 * Fixed a bug that caused `resultAt` to give wrong answers most of the time.
+* Changed `resultAt` to an `IndexedLens` and moved it to `Control.Lens.IndexedLens`
 
 3.2
 ---